[Day 10] Unsatisfying solution for part 2.
I have mostly solved this, but some problems with the method and the 'S'
position that I am using that causes 1 fewer blocks to be inside than
should be. I am compensating for the off-by-one at the end, which gives
the correct answer, but the code does not handle this case. I would like
to try to re-write with an alternate method which is more tolerant to
the situations in my input data.

Also fixed a few other bugs and formatting issues too.