[Day 16] Solution to part 1.
A solution to part 1 of day 16. Not too bad. Fairly proud of the mirror logic here today, as I was about to descend into dealing with cases. Started recursive, but then realised I needed to keep track of the branches due to splitters, so implemented a stack for the beams instead. Seems to work now.