[Day 20] Solution for part 1 and 2.
This was a bit of a long one, and the implementation took a while. A few dead ends on this too, but got the right answer in the end. Thanks to HyperNeutrino about the suggestuion of the structure of the input. Mine followed the same, so was able to use a similar method.