[Day 23] Solution to part 1.
Now I have some free time, I am catching up on the last few days. Here's
a solution for part 1 of day 23. It took me a while to work out how the
recursive DFS was going to work in rust, but passing the mutable borrow
to the original object worked in the end.